[CPDLP-2073] Amend scripts to develop statements information for 2023 cohort
