+++ db/data/statements/ecf.csv
@@ -1,63 +1,95 @@
-name,cohort,deadline_date,payment_date,contract_version,output_fee
+name,cohort,deadline_date,payment_date,output_fee
 
-November 2021,2021,2021-11-30,2021-11-30,0.0.1,true
-January 2022,2021,2021-12-31,2022-1-25,0.0.1,false
-February 2022,2021,2022-1-31,2022-2-25,0.0.1,true
-March 2022,2021,2022-2-28,2022-3-25,0.0.1,false
-April 2022,2021,2022-3-31,2022-4-25,0.0.1,false
-May 2022,2021,2022-4-30,2022-5-25,0.0.1,true
-June 2022,2021,2022-5-31,2022-6-25,0.0.1,true
-July 2022,2021,2022-6-30,2022-7-25,0.0.1,false
-August 2022,2021,2022-7-31,2022-8-25,0.0.1,false
-September 2022,2021,2022-8-31,2022-9-25,0.0.1,false
-October 2022,2021,2022-9-30,2022-10-25,0.0.1,true
-November 2022,2021,2022-10-31,2022-11-25,0.0.1,false
-December 2022,2021,2022-11-30,2022-12-25,0.0.1,false
-January 2023,2021,2023-12-31,2023-1-25,0.0.1,false
-February 2023,2021,2023-1-31,2023-2-25,0.0.1,true
-March 2023,2021,2023-2-28,2023-3-25,0.0.1,false
-April 2023,2021,2023-3-31,2023-4-25,0.0.1,false
-May 2023,2021,2023-4-30,2023-5-25,0.0.1,true
-June 2023,2021,2023-5-31,2023-6-25,0.0.1,false
-July 2023,2021,2023-6-30,2023-7-25,0.0.1,false
-August 2023,2021,2023-7-31,2023-8-25,0.0.1,false
-September 2023,2021,2023-8-31,2023-9-25,0.0.1,false
-October 2023,2021,2023-9-30,2023-10-25,0.0.1,true
-November 2023,2021,2023-10-31,2023-11-25,0.0.1,false
-December 2023,2021,2023-11-30,2023-12-25,0.0.1,false
-January 2024,2021,2024-12-31,2024-1-25,0.0.1,false
-February 2024,2021,2024-1-31,2024-2-25,0.0.1,true
-March 2024,2021,2024-2-28,2024-3-25,0.0.1,false
-April 2024,2021,2024-3-31,2024-4-25,0.0.1,false
-May 2024,2021,2024-4-30,2024-5-25,0.0.1,true
+November 2021,2021,2021-11-30,2021-11-30,true
+January 2022,2021,2021-12-31,2022-1-25,false
+February 2022,2021,2022-1-31,2022-2-25,true
+March 2022,2021,2022-2-28,2022-3-25,false
+April 2022,2021,2022-3-31,2022-4-25,false
+May 2022,2021,2022-4-30,2022-5-25,true
+June 2022,2021,2022-5-31,2022-6-25,true
+July 2022,2021,2022-6-30,2022-7-25,false
+August 2022,2021,2022-7-31,2022-8-25,false
+September 2022,2021,2022-8-31,2022-9-25,false
+October 2022,2021,2022-9-30,2022-10-25,true
+November 2022,2021,2022-10-31,2022-11-25,false
+December 2022,2021,2022-11-30,2022-12-25,false
+January 2023,2021,2023-12-31,2023-1-25,false
+February 2023,2021,2023-1-31,2023-2-25,true
+March 2023,2021,2023-2-28,2023-3-25,false
+April 2023,2021,2023-3-31,2023-4-25,false
+May 2023,2021,2023-4-30,2023-5-25,true
+June 2023,2021,2023-5-31,2023-6-25,false
+July 2023,2021,2023-6-30,2023-7-25,false
+August 2023,2021,2023-7-31,2023-8-25,false
+September 2023,2021,2023-8-31,2023-9-25,false
+October 2023,2021,2023-9-30,2023-10-25,true
+November 2023,2021,2023-10-31,2023-11-25,false
+December 2023,2021,2023-11-30,2023-12-25,false
+January 2024,2021,2024-12-31,2024-1-25,false
+February 2024,2021,2024-1-31,2024-2-25,true
+March 2024,2021,2024-2-28,2024-3-25,false
+April 2024,2021,2024-3-31,2024-4-25,false
+May 2024,2021,2024-4-30,2024-5-25,true
 
-November 2022,2022,2022-11-30,2022-11-30,0.0.1,true
-January 2023,2022,2022-12-31,2023-1-25,0.0.1,false
-February 2023,2022,2023-1-31,2023-2-25,0.0.1,true
-March 2023,2022,2023-2-28,2023-3-25,0.0.1,false
-April 2023,2022,2023-3-31,2023-4-25,0.0.1,false
-May 2023,2022,2023-4-30,2023-5-25,0.0.1,true
-June 2023,2022,2023-5-31,2023-6-25,0.0.1,true
-July 2023,2022,2023-6-30,2023-7-25,0.0.1,false
-August 2023,2022,2023-7-31,2023-8-25,0.0.1,false
-September 2023,2022,2023-8-31,2023-9-25,0.0.1,false
-October 2023,2022,2023-9-30,2023-10-25,0.0.1,true
-November 2023,2022,2023-10-31,2023-11-25,0.0.1,false
-December 2023,2022,2023-11-30,2023-12-25,0.0.1,false
-January 2024,2022,2024-12-31,2024-1-25,0.0.1,false
-February 2024,2022,2024-1-31,2024-2-25,0.0.1,true
-March 2024,2022,2024-2-28,2024-3-25,0.0.1,false
-April 2024,2022,2024-3-31,2024-4-25,0.0.1,false
-May 2024,2022,2024-4-30,2024-5-25,0.0.1,true
-June 2024,2022,2024-5-31,2024-6-25,0.0.1,false
-July 2024,2022,2024-6-30,2024-7-25,0.0.1,false
-August 2024,2022,2024-7-31,2024-8-25,0.0.1,false
-September 2024,2022,2024-8-31,2024-9-25,0.0.1,false
-October 2024,2022,2024-9-30,2024-10-25,0.0.1,true
-November 2024,2022,2024-10-31,2024-11-25,0.0.1,false
-December 2024,2022,2024-11-30,2024-12-25,0.0.1,false
-January 2025,2022,2025-12-31,2025-1-25,0.0.1,false
-February 2025,2022,2025-1-31,2025-2-25,0.0.1,true
-March 2025,2022,2025-2-28,2025-3-25,0.0.1,false
-April 2025,2022,2025-3-31,2025-4-25,0.0.1,false
-May 2025,2022,2025-4-30,2025-5-25,0.0.1,true
+November 2022,2022,2022-11-30,2022-11-30,true
+January 2023,2022,2022-12-31,2023-1-25,false
+February 2023,2022,2023-1-31,2023-2-25,true
+March 2023,2022,2023-2-28,2023-3-25,false
+April 2023,2022,2023-3-31,2023-4-25,false
+May 2023,2022,2023-4-30,2023-5-25,true
+June 2023,2022,2023-5-31,2023-6-25,true
+July 2023,2022,2023-6-30,2023-7-25,false
+August 2023,2022,2023-7-31,2023-8-25,false
+September 2023,2022,2023-8-31,2023-9-25,false
+October 2023,2022,2023-9-30,2023-10-25,true
+November 2023,2022,2023-10-31,2023-11-25,false
+December 2023,2022,2023-11-30,2023-12-25,false
+January 2024,2022,2024-12-31,2024-1-25,false
+February 2024,2022,2024-1-31,2024-2-25,true
+March 2024,2022,2024-2-28,2024-3-25,false
+April 2024,2022,2024-3-31,2024-4-25,false
+May 2024,2022,2024-4-30,2024-5-25,true
+June 2024,2022,2024-5-31,2024-6-25,false
+July 2024,2022,2024-6-30,2024-7-25,false
+August 2024,2022,2024-7-31,2024-8-25,false
+September 2024,2022,2024-8-31,2024-9-25,false
+October 2024,2022,2024-9-30,2024-10-25,true
+November 2024,2022,2024-10-31,2024-11-25,false
+December 2024,2022,2024-11-30,2024-12-25,false
+January 2025,2022,2025-12-31,2025-1-25,false
+February 2025,2022,2025-1-31,2025-2-25,true
+March 2025,2022,2025-2-28,2025-3-25,false
+April 2025,2022,2025-3-31,2025-4-25,false
+May 2025,2022,2025-4-30,2025-5-25,true
+
+November 2023,2023,2023-10-31,2023-11-25,false
+December 2023,2023,2023-11-30,2023-12-25,false
+January 2024,2023,2023-12-31,2024-1-25,true
+February 2024,2023,2024-1-31,2024-2-25,false
+March 2024,2023,2024-2-29,2024-3-25,false
+April 2024,2023,2024-3-31,2024-4-25,true
+May 2024,2023,2024-4-30,2024-5-25,false
+June 2024,2023,2024-5-31,2024-6-25,false
+July 2024,2023,2024-6-30,2024-7-25,false
+August 2024,2023,2024-7-31,2024-8-25,true
+September 2024,2023,2024-8-31,2024-9-25,false
+October 2024,2023,2024-9-30,2024-10-25,true
+November 2024,2023,2024-10-31,2024-11-25,false
+December 2024,2023,2024-11-30,2024-12-25,false
+January 2025,2023,2025-12-31,2025-1-25,true
+February 2025,2023,2025-1-31,2025-2-25,false
+March 2025,2023,2025-2-28,2025-3-25,false
+April 2025,2023,2025-3-31,2025-4-25,true
+May 2025,2023,2025-4-30,2025-5-25,false
+June 2025,2023,2025-5-31,2025-6-25,false
+July 2025,2023,2025-6-30,2025-7-25,false
+August 2025,2023,2025-7-31,2025-8-25,true
+September 2025,2023,2025-8-31,2025-9-25,false
+October 2025,2023,2025-9-30,2025-10-25,true
+November 2025,2023,2025-10-31,2025-11-25,false
+December 2025,2023,2025-11-30,2025-12-25,false
+January 2026,2023,2026-12-31,2026-1-25,true
+February 2026,2023,2026-1-31,2026-2-25,false
+March 2026,2023,2026-2-28,2026-3-25,false
+April 2026,2023,2026-3-31,2026-4-25,true
+May 2026,2023,2026-4-30,2026-5-25,false
