+++ db/legacy_seeds/initial_seed.rb
@@ -85,4 +85,4 @@ NPQLeadProvider.all.each do |lp|
   lp.update!(cpd_lead_provider: CpdLeadProvider.find_by(name: lp.name))
 end
 
-Importers::SeedStatements.new.call
+Importers::CreateStatements.new.call
