--- db/new_seeds/base/add_seed_statements.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Importers::SeedStatements.new.call
