+++ db/new_seeds/base/add_statements.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+Importers::CreateStatements.new.call
