+++ spec/features/scenarios/onboarding_a_withdrawn_participant/cip_to_fip_spec.rb
@@ -81,7 +81,7 @@ RSpec.feature "CIP to FIP - Onboarding a withdrawn participant",
 
         travel_to(milestone_started.milestone_date - 2.months) do
           Seeds::CallOffContracts.new.call
-          Importers::SeedStatements.new.call
+          Importers::CreateStatements.new.call
         end
 
         and_sit_at_pupil_premium_school_reported_programme "Original SIT", "CIP"
