+++ spec/features/scenarios/transfering_a_participant/cip_to_cip_spec.rb
@@ -79,7 +79,7 @@ RSpec.feature "CIP to CIP - Transfer a participant",
 
         travel_to(milestone_started.milestone_date - 2.months) do
           Seeds::CallOffContracts.new.call
-          Importers::SeedStatements.new.call
+          Importers::CreateStatements.new.call
         end
 
         and_sit_at_pupil_premium_school_reported_programme "Original SIT", "CIP"
