+++ spec/features/scenarios/transfering_a_participant/fip_to_fip_with_different_provider_spec.rb
@@ -81,7 +81,7 @@ RSpec.feature "FIP to FIP with different provider - Transfer a participant",
 
         travel_to(milestone_started.milestone_date - 2.months) do
           Seeds::CallOffContracts.new.call
-          Importers::SeedStatements.new.call
+          Importers::CreateStatements.new.call
         end
 
         and_sit_at_pupil_premium_school_reported_programme "Original SIT", "FIP"
